converting

+++ process-single.sh
@@ -7,17 +7,22 @@ CODE=$2
 COUNTRY=$3
 
 echo "---- ${CONTINENT} ${CODE} ${COUNTRY} ----"
-curl -o temp/${COUNTRY}.osm.pbf https://download.geofabrik.de/${CONTINENT}/${COUNTRY}-latest.osm.pbf
+if [ -f temp/${COUNTRY}.geo.jsonseq.bz2 ]; then
+        echo "Already done"
+        exit 0
+fi
+
+if [ ! -f  temp/${COUNTRY}.osm.pbf  ]; then
+        curl -o temp/${COUNTRY}.osm.pbf https://download.geofabrik.de/${CONTINENT}/${COUNTRY}-latest.osm.pbf
+fi
 
 echo "Converting PBF to GeoJson"
 # https://docs.osmcode.org/osmium/latest/osmium-export.html
-osmium export --add-unique-id=type_id --overwrite --config=osmium-config.json --output="temp/${COUNTRY}.geo.jsonseq" "temp/${COUNTRY}.osm.pbf"
+osmium export --add-unique-id=type_id --overwrite --config=osmium-config.json --output-format=geojsonseq --output="temp/${COUNTRY}.geo.jsonseq" "temp/${COUNTRY}.osm.pbf"
 echo "Rows: `wc -l temp/${COUNTRY}.geo.jsonseq`"
 
 echo "Compressing..."
-#time gzip --force -9 -k temp/${COUNTRY}.geo.jsonseq
-#time zstd --force -15 temp/${COUNTRY}.geo.jsonseq
-time pbzip2 temp/${COUNTRY}.geo.jsonseq
+time pbzip2 -f -k temp/${COUNTRY}.geo.jsonseq
 ls -lh temp/${COUNTRY}.*
 
 echo "Upload..."
